Start reworking the crew dashboard
