+++ resources/views/components/sidemenus/content-moderator.blade.php
@@ -7,28 +7,28 @@
             :icon="'M2.25 21h19.5m-18-18v18m10.5-18v18m6-13.5V21M6.75 6.75h.75m-.75 3h.75m-.75 3h.75m3-6h.75m-.75 3h.75m-.75 3h.75M6.75 21v-3.375c0-.621.504-1.125 1.125-1.125h2.25c.621 0 1.125.504 1.125 1.125V21M3 3h12m-.75 4.5H21m-3.75 3.75h.008v.008h-.008v-.008zm0 3h.008v.008h-.008v-.008zm0 3h.008v.008h-.008v-.008z'"
             :param="''"
             :roleColour="'crew'"
-            :badgeText="\App\Models\Team::awaitingApproval()->count()"></x-sidebar-link-content-mod>
+            :badgeText="\App\Models\Company::where('is_approved', 0)->count()"></x-sidebar-link-content-mod>
         <x-sidebar-link-content-mod
             :label="'Presentations'"
             :route="'moderator.presentations.index'"
             :icon="'M3.75 3v11.25A2.25 2.25 0 006 16.5h2.25M3.75 3h-1.5m1.5 0h16.5m0 0h1.5m-1.5 0v11.25A2.25 2.25 0 0118 16.5h-2.25m-7.5 0h7.5m-7.5 0l-1 3m8.5-3l1 3m0 0l.5 1.5m-.5-1.5h-9.5m0 0l-.5 1.5M9 11.25v1.5M12 9v3.75m3-6v6'"
             :param="''"
             :roleColour="'crew'"
-            :badgeText="\App\Models\Presentation::awaitingApproval()->count()"></x-sidebar-link-content-mod>
+            :badgeText="\App\Models\Presentation::where('is_approved', 0)->count()"></x-sidebar-link-content-mod>
         <x-sidebar-link-content-mod
             :label="'Booths'"
             :route="'moderator.booths.index'"
             :icon="'M13.5 21v-7.5a.75.75 0 01.75-.75h3a.75.75 0 01.75.75V21m-4.5 0H2.36m11.14 0H18m0 0h3.64m-1.39 0V9.349m-16.5 11.65V9.35m0 0a3.001 3.001 0 003.75-.615A2.993 2.993 0 009.75 9.75c.896 0 1.7-.393 2.25-1.016a2.993 2.993 0 002.25 1.016c.896 0 1.7-.393 2.25-1.016a3.001 3.001 0 003.75.614m-16.5 0a3.004 3.004 0 01-.621-4.72L4.318 3.44A1.5 1.5 0 015.378 3h13.243a1.5 1.5 0 011.06.44l1.19 1.189a3 3 0 01-.621 4.72m-13.5 8.65h3.75a.75.75 0 00.75-.75V13.5a.75.75 0 00-.75-.75H6.75a.75.75 0 00-.75.75v3.75c0 .415.336.75.75.75z'"
             :param="''"
             :roleColour="'crew'"
-            :badgeText="\App\Models\Booth::awaitingApproval()->count()"></x-sidebar-link-content-mod>
+            :badgeText="\App\Models\Booth::where('is_approved', 0)->count()"></x-sidebar-link-content-mod>
         <x-sidebar-link-content-mod
             :label="'Sponsorships'"
             :route="'moderator.sponsors.index'"
             :icon="'M21 11.25v8.25a1.5 1.5 0 01-1.5 1.5H5.25a1.5 1.5 0 01-1.5-1.5v-8.25M12 4.875A2.625 2.625 0 109.375 7.5H12m0-2.625V7.5m0-2.625A2.625 2.625 0 1114.625 7.5H12m0 0V21m-8.625-9.75h18c.621 0 1.125-.504 1.125-1.125v-1.5c0-.621-.504-1.125-1.125-1.125h-18c-.621 0-1.125.504-1.125 1.125v1.5c0 .621.504 1.125 1.125 1.125z'"
             :param="''"
             :roleColour="'crew'"
-            :badgeText="\App\Models\SponsorTier::awaitingApproval()->count()"></x-sidebar-link-content-mod>
+            :badgeText="\App\Models\Company::where('is_sponsorship_approved', 0)->count()"></x-sidebar-link-content-mod>
         <x-sidebar-link-content-mod
             :label="'Users'"
             :route="'moderator.users.index'"
